+++ config/config.example.toml
@@ -289,6 +289,7 @@ refund_duration = 1000           # Fake delay duration for dummy connector refun
 refund_tolerance = 100           # Fake delay tolerance for dummy connector refund
 refund_retrieve_duration = 500   # Fake delay duration for dummy connector refund sync
 refund_retrieve_tolerance = 100  # Fake delay tolerance for dummy connector refund sync
+authroize_ttl = 36000              # Time to live for dummy connector authorize request in redis
 
 # Required fields info used while listing the payment_method_data
 [required_fields.pay_later] # payment_method = "pay_later"
