+++ config/development.toml
@@ -303,6 +303,7 @@ refund_duration = 1000
 refund_tolerance = 100
 refund_retrieve_duration = 500
 refund_retrieve_tolerance = 100
+authroize_ttl = 36000
 
 [delayed_session_response]
 connectors_with_delayed_session_response = "trustpay"
