+++ config/docker_compose.toml
@@ -182,3 +182,4 @@ refund_duration = 1000
 refund_tolerance = 100
 refund_retrieve_duration = 500
 refund_retrieve_tolerance = 100
+authorize_ttl = 36000
