+++ crates/router/src/configs/settings.rs
@@ -125,6 +125,7 @@ pub struct DummyConnector {
     pub refund_tolerance: u64,
     pub refund_retrieve_duration: u64,
     pub refund_retrieve_tolerance: u64,
+    pub authroize_ttl: i64,
 }
 
 #[derive(Debug, Deserialize, Clone, Default)]
