+++ crates/router/src/routes/dummy_connector/errors.rs
@@ -34,6 +34,9 @@ pub enum DummyConnectorErrors {
 
     #[error(error_type = ErrorType::ServerNotAvailable, code = "DC_07", message = "Error occurred while storing the payment")]
     PaymentStoringError,
+
+    #[error(error_type = ErrorType::InvalidRequestError, code = "DC_08", message = "Payment declined")]
+    PaymentDeclined,
 }
 
 impl core::fmt::Display for DummyConnectorErrors {
@@ -76,6 +79,9 @@ impl common_utils::errors::ErrorSwitch<api_models::errors::types::ApiErrorRespon
             }
             Self::PaymentStoringError => {
                 AER::InternalServerError(ApiError::new("DC", 7, self.error_message(), None))
+            }
+            Self::PaymentDeclined => {
+                AER::BadRequest(ApiError::new("DC", 8, self.error_message(), None))
             }
         }
     }
